// backend/src/db/schemas/ai-mcp-server-tools.ts
// Code generated by automation script, DO NOT EDIT.
// Automated by pulling database and generating zod schema
// To update. Just run npm run generate:schema
// Written by akhilmhdh.

import { z } from "zod";

import { TImmutableDBKeys } from "./models";

export const AiMcpServerToolsSchema = z.object({
  id: z.string().uuid(),
  name: z.string(),
  description: z.string().nullable().optional(),
  inputSchema: z.unknown().nullable().optional(),
  aiMcpServerId: z.string().uuid()
});

export type TAiMcpServerTools = z.infer<typeof AiMcpServerToolsSchema>;
export type TAiMcpServerToolsInsert = Omit<z.input<typeof AiMcpServerToolsSchema>, TImmutableDBKeys>;
export type TAiMcpServerToolsUpdate = Partial<Omit<z.input<typeof AiMcpServerToolsSchema>, TImmutableDBKeys>>;